pyproject.toml
[project]
name = "celeste-client"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.1.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
